+++ system/platform-core/src/main/java/org/platformlambda/core/system/EventEmitter.java
@@ -38,6 +38,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class EventEmitter {
     private static final Logger log = LoggerFactory.getLogger(EventEmitter.class);
@@ -83,6 +84,8 @@ public class EventEmitter {
     private boolean multicastEnabled = false;
     private boolean journalEnabled = false;
     private boolean eventHttpEnabled = false;
+    private static final AtomicInteger warmUpCounter = new AtomicInteger(0);
+    private static boolean warm = false;
     private static final EventEmitter INSTANCE = new EventEmitter();
 
     private EventEmitter() {
@@ -721,7 +724,7 @@ public class EventEmitter {
                 /*
                  * The "event.script.manager" and "task.executor" are reserved function route names
                  * for Event Script. The system will run them directly using virtual threads for
-                 * performance optimization.
+                 * performance optimization after warming up the virtual thread system.
                  *
                  * Since these two functions are event routers themselves, functional isolation,
                  * serialization and I/O immutability are guaranteed.
@@ -729,8 +732,17 @@ public class EventEmitter {
                  * Therefore, it is safe to eliminate additional processing overheads.
                  */
                 if (TASK_EXECUTOR.equals(route) || EVENT_MANAGER.equals(route)) {
-                    Platform.getInstance().getVirtualThreadExecutor().submit(() ->
-                            runTaskExecutor(out, target.getManager().getService().getFunction()));
+                    if (warm) {
+                        Platform.getInstance().getVirtualThreadExecutor().submit(() ->
+                                runTaskExecutor(out, target.getManager().getService().getFunction()));
+                    } else {
+                        system.send(route, out.toBytes());
+                        var counter = warmUpCounter.incrementAndGet();
+                        if (counter > 8) {
+                            warm = true;
+                            log.info("Virtual thread optimization completed");
+                        }
+                    }
                 } else {
                     system.send(route, out.toBytes());
                 }